Handle 400 errors when creating a new customer
I wrote a test, test_create_customer, that includes two requests:
 - Valid customer data
 - Invalid customer data

The Dubai API responds with a 400 error when an attribute is missing from customer data.  We
raise an exception, SoftixException, for any response that is not a 200.

The betamax recorder is reconfigured to match on body, instead of json_body.  Although I
prefer to the latter, the Dubai API is inconsistent.  Generating a token does not accept
JSON, while ALL other calls do.