Validate required customer attrs and character len
Check if nationality and countrycode are two characters long and raise
an exception.  John Bender was unable to create a new customer and the
error, returned from API, was uninformative: "Not all customer details
were provided or some details were incorrect."

Added a validation step that checks if all the required fields, based
off of our testing, are included in the customer request. Raise
MissingRequiredCustomerField if any field is missing.